+++ app/services/cfe_service.rb
@@ -3,16 +3,16 @@ class CfeService
     def call(session_data)
       cfe_connection = CfeConnection.connection
       cfe_assessment_id = create_assessment_id(cfe_connection, session_data)
-      SubmitDependantsService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitProceedingsService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitEmploymentIncomeService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitBenefitsService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitIrregularIncomeService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitVehicleService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitAssetsService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitRegularTransactionsService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitApplicantService.call(cfe_connection, cfe_assessment_id, session_data)
-      SubmitPartnerService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitDependantsService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitProceedingsService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitEmploymentIncomeService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitBenefitsService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitIrregularIncomeService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitVehicleService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitAssetsService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitRegularTransactionsService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitApplicantService.call(cfe_connection, cfe_assessment_id, session_data)
+      Cfe::SubmitPartnerService.call(cfe_connection, cfe_assessment_id, session_data)
       result(cfe_connection, cfe_assessment_id, session_data)
     end
